backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
django-filter>=23.0,<25.0
psycopg2-binary>=2.9,<3.0
gunicorn>=21.0,<23.0
Pillow>=10.0,<11.0
